Moving server side code to a separate directory
