+++ bin/www
@@ -7,7 +7,7 @@
 const app = require('../app');
 const debug = require('debug')('server:server');
 const http = require('http');
-const config = require('../config/config');
+const config = require('../server/config/config');
 /**
  * Get port from environment and store in Express.
  */
