--- public/v1/readme.txt
@@ -1,1 +1,0 @@
-This is a placeholder.
