+++ server/controllers/company.controller.js
@@ -1,4 +1,4 @@
-const { Company } = require('../models');
+const { Company } = require('../models/index');
 const { to, reE, reS } = require('../services/util.service');
 
 const create = async(req, res) => {
