+++ server/elasticsearch/indexmanager.js
